+++ Zilon.Core/Zilon.Core/Persons/MonsterPerson.cs
@@ -13,7 +13,7 @@ namespace Zilon.Core.Persons
     public class MonsterPerson : IPerson
     {
         public int Id { get; set; }
-        public float Hp { get; }
+        public int Hp { get; }
         public IEquipmentCarrier EquipmentCarrier => throw new NotSupportedException("Для монстров не поддерживается явная экипировка");
 
         public ITacticalActCarrier TacticalActCarrier { get; }
